+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
 ## [0.7.0] - 2022-11-07
 
 - [DemoApp/SDK] #113 Add theme management and create Orange and InnovationCup themes